+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Updated Page size button to prevent overflow (#62)
 - Added reset table state on tab change (#67)
 - Updated Table Editor to improve UI/UX (#66)
+- Updated Sorting options (#69)
 
 ## 1.2.0 (2023-09-05)